Create more flexible NN class, explore configs, and gain classification insight
Create a subclass of Neural_Network with an init method that allows
supports modified sizes for the input, hidden and output layers.

Exlore input layer with bias component and larger hidden layer to
see if that helps network learn "add one" counting.  Realize
through exploration that the output is constrained by sigmoid function
so will always be in range 0-1. The "add one" is not a classifier
so the neural network will fail to learn this.

Lay out plan to build a descrimintive network for a "count to 10"
or "what number comes next" sequence prediction network.